Koha How-to template revisions: CSS, markup, text edits
This patch makes changes to the how-to's CSS, markup and text.

- Markup has been corrected and expanded to allow for better custom
  styling.
- CSS has been added to give the how-to a distinct style.
- Some spelling and grammar error have been corrected. Other text
  strings have been modified for clarity.

To test:

Apply the patch and step through each step of the how-to. The style
should look clear and consistent. The functionality of the tuorial
should be unchanged.

Note: This patch contains a lot of whitespace changes, so please ignore
whitespace when looking at the diff.